refactor: 주문 관련 코드 제거
