+++ src/main/java/woowacourse/shoppingcart/application/ProductService.java
@@ -4,8 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.ProductRequest;
-import woowacourse.shoppingcart.exception.InvalidProductException;
+import woowacourse.shoppingcart.dto.product.ProductRequest;
 import woowacourse.shoppingcart.exception.NotFoundProductException;
 
 import java.util.List;
