+++ src/main/java/woowacourse/shoppingcart/dto/cart/CartItemAddRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cart;
 
 public class CartItemAddRequest {
     private Long productId;
