+++ src/main/java/woowacourse/shoppingcart/dto/cart/CartItemsResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cart;
 
 import woowacourse.shoppingcart.domain.Cart;
