+++ src/main/java/woowacourse/shoppingcart/dto/customer/CustomerResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.customer;
 
 public class CustomerResponse {
     private String email;
