+++ src/main/java/woowacourse/shoppingcart/dto/customer/SignUpRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.customer;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.Pattern;
