+++ src/main/resources/schema.sql
@@ -46,30 +46,30 @@ alter table cart_item
     add constraint fk_cart_item_to_product
         foreign key (product_id) references product (id);
 
-create table orders
-(
-    id          bigint not null auto_increment,
-    customer_id bigint not null,
-    primary key (id)
-) engine=InnoDB default charset=utf8mb4;
-
-alter table orders
-    add constraint fk_orders_to_customer
-        foreign key (customer_id) references customer (id);
-
-create table orders_detail
-(
-    id         bigint  not null auto_increment,
-    orders_id  bigint  not null,
-    product_id bigint  not null,
-    quantity   integer not null,
-    primary key (id)
-) engine=InnoDB default charset=utf8mb4;
-
-alter table orders_detail
-    add constraint fk_orders_detail_to_orders
-        foreign key (orders_id) references orders (id);
-
-alter table orders_detail
-    add constraint fk_orders_detail_to_product
-        foreign key (product_id) references product (id);
+-- create table orders
+-- (
+--     id          bigint not null auto_increment,
+--     customer_id bigint not null,
+--     primary key (id)
+-- ) engine=InnoDB default charset=utf8mb4;
+--
+-- alter table orders
+--     add constraint fk_orders_to_customer
+--         foreign key (customer_id) references customer (id);
+--
+-- create table orders_detail
+-- (
+--     id         bigint  not null auto_increment,
+--     orders_id  bigint  not null,
+--     product_id bigint  not null,
+--     quantity   integer not null,
+--     primary key (id)
+-- ) engine=InnoDB default charset=utf8mb4;
+--
+-- alter table orders_detail
+--     add constraint fk_orders_detail_to_orders
+--         foreign key (orders_id) references orders (id);
+--
+-- alter table orders_detail
+--     add constraint fk_orders_detail_to_product
+--         foreign key (product_id) references product (id);
